+++ src/tdme/tools/installer/Installer.cpp
@@ -3,8 +3,10 @@
 #include <stdlib.h>
 #include <stdio.h>
 
+#include <algorithm>
 #include <string>
 #include <unordered_map>
+#include <unordered_set>
 #include <vector>
 
 #include <tdme/application/Application.h>
@@ -48,9 +50,14 @@
 
 using tdme::tools::installer::Installer;
 
+using std::distance;
+using std::reverse;
+using std::sort;
 using std::string;
 using std::to_string;
+using std::unique;
 using std::unordered_map;
+using std::unordered_set;
 using std::vector;
 
 using tdme::application::Application;
@@ -698,11 +705,16 @@ void Installer::onActionPerformed(GUIActionListener_Type* type, GUIElementNode* 
 						dynamic_cast<GUIElementNode*>(installer->engine->getGUI()->getScreen("installer_uninstalling")->getNodeById("progressbar"))->getController()->setValue(MutableString(0.0f, 2));
 						installer->installThreadMutex.unlock();
 
-						//
+						// remove files that we installed
 						auto hadException = false;
 						vector<string> log;
 						try {
 							FileSystem::getStandardFileSystem()->getContentAsStringArray(".", "install.files.db", log);
+						} catch (Exception& exception) {
+							installer->popUps->getInfoDialogScreenController()->show("An error occurred:", exception.what());
+							hadException = true;
+						}
+						if (hadException == false) {
 							for (auto i = 1; i < log.size(); i++) {
 								try {
 									installer->installThreadMutex.lock();
@@ -718,19 +730,45 @@ void Installer::onActionPerformed(GUIActionListener_Type* type, GUIElementNode* 
 									Console::println(string("UninstallThread::run(): An error occurred: ") + innerException.what());
 								}
 							}
-						} catch (Exception& exception) {
-							installer->popUps->getInfoDialogScreenController()->show("An error occurred:", exception.what());
-							hadException = true;
 						}
 
-						try {
-							FileSystem::getStandardFileSystem()->removeFile(".", "install.files.db");
-							FileSystem::getStandardFileSystem()->removeFile(".", "install.components.db");
-							FileSystem::getStandardFileSystem()->removeFile(".", "install.version.db");
-							FileSystem::getStandardFileSystem()->removeFile(".", "console.log");
-						} catch (Exception& exception) {
-							installer->popUps->getInfoDialogScreenController()->show("An error occurred:", exception.what());
-							hadException = true;
+						// remove folders that we created non recursive
+						if (hadException == false) {
+							auto installFolder = log[0] + "/";
+							vector<string> folders;
+							for (auto i = 1; i < log.size(); i++) {
+								auto folderCandidate = FileSystem::getStandardFileSystem()->getPathName(log[i]);
+								if (folderCandidate.empty() == true) continue;
+								if (StringUtils::startsWith(folderCandidate, installFolder) == true &&
+									FileSystem::getStandardFileSystem()->isPath(folderCandidate) == true) folders.push_back(folderCandidate);
+							}
+							sort(folders.begin(), folders.end());
+							reverse(folders.begin(), folders.end());
+							auto newEnd = unique(folders.begin(), folders.end());
+							folders.resize(distance(folders.begin(), newEnd));
+							for (auto folder: folders) {
+								try {
+									FileSystem::getStandardFileSystem()->removePath(
+										folder,
+										false
+									);
+								} catch (Exception& innerException) {
+									Console::println(string("UninstallThread::run(): An error occurred: ") + innerException.what());
+								}
+							}
+						}
+
+						// remove install databases and console.log
+						if (hadException == false) {
+							try {
+								FileSystem::getStandardFileSystem()->removeFile(".", "install.files.db");
+								FileSystem::getStandardFileSystem()->removeFile(".", "install.components.db");
+								FileSystem::getStandardFileSystem()->removeFile(".", "install.version.db");
+								FileSystem::getStandardFileSystem()->removeFile(".", "console.log");
+							} catch (Exception& exception) {
+								installer->popUps->getInfoDialogScreenController()->show("An error occurred:", exception.what());
+								hadException = true;
+							}
 						}
 
 						//
